requirements.txt
google-api-python-client==2.105.0
google-cloud-aiplatform==1.35.0
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.0.324
scikit-learn==1.3.2
fastapi==0.104.0
Jinja2==3.1.2
python-multipart==0.0.6
uvicorn==0.23.2